requirements.txt
catboost==1.2.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
graphviz==0.20.1
importlib-resources==6.1.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
plotly==5.17.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
tenacity==8.2.3
threadpoolctl==3.2.0
tzdata==2023.3
zipp==3.17.0
